msgpackconverter: Fix buffer reuse for encoding JSON strings
The `dst` given always has len 0, we should've looked at the cap() to
see how much we can append to it without reallocating.

While at it, if the buffer has space for 2+len(input), we might just get
lucky if nothing needs escaping so try that. Otherwise, we prefer
append() to decide a new size anyway.